Revert "🚚 move actions into *.action files"
This reverts commit d47792f93094ab1ec31a5954dfdf6e83ebed09da.
