+++ src/app/(routes)/(app)/settings/account/page.tsx
@@ -5,10 +5,7 @@ import { UpdateUserInfosForm } from "~/app/(components)/update-user-infos-form";
 import { Button } from "~/app/(components)/button";
 
 // utils
-import {
-  getAuthedUser,
-  redirectIfNotAuthed
-} from "~/app/(actions)/auth.action";
+import { getAuthedUser, redirectIfNotAuthed } from "~/app/(actions)/auth";
 import { updateUserProfileInfosInputValidator } from "~/app/(models)/dto/update-profile-info-input-validator";
 
 // types
