+++ src/app/(routes)/api/auth/callback/route.ts
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
-import { loginUser } from "~/app/(actions)/auth.action";
+import { loginUser } from "~/app/(actions)/auth";
 import { env } from "~/env";
 import { isValidURLPathname } from "~/lib/shared/utils.shared";
 import type { NextRequest } from "next/server";
